kitchen_agent/requirements.txt
google-adk>=1.7.0
google-adk[a2a]>=1.8.0
google-generativeai>=1.26.0